Added ajax for live update
On the status page.
